moved bayes and state_translate to data_pipeline/
